shared-models/pyproject.toml
[project]
name = "partner-agent-shared-models"
version = "0.1.3"
description = "Shared models and schemas for partner agent integration"
authors = [
    {name = "Partner Agent Team", email = "team@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.0",
    "psycopg-pool>=3.2.0",
    "alembic>=1.13.0",
    "structlog>=23.2.0",
    "fastapi>=0.104.0,<0.117.0",
    "pydantic>=2.5.0",
    "cloudevents>=1.10.0",
    "httpx>=0.25.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "bcrypt>=4.0.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.2.1",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=5.13.0",
    "mypy>=1.17.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shared_models"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
# Note: We use StructuredLogger Protocol in shared_models.logging to allow
# structured logging with arbitrary keyword arguments without type errors

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.2.1",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=5.13.0",
    "mypy>=1.17.1",
]
